Create wc-redirect-category-to-single-page.php
This code checks if the current page is a product category page. If it is, it retrieves the category's term ID and counts the number of products in that category. If the category has only one product, it retrieves the ID of that product and redirects the user to its single product page.